Moved ShellyPowerData and ShellyThermostatData to SHaredContracts

+++ ShellyConnector/ShellyConnector/DataContracts/ShellyEMeterData.cs
@@ -8,13 +8,13 @@ namespace ShellyConnector.DataContracts
 {
     public class Emeter
     {
-        public double Power { get; set; }
-        public double Pf { get; set; }
-        public double Current { get; set; }
-        public double Voltage { get; set; }
+        public decimal Power { get; set; }
+        public decimal Pf { get; set; }
+        public decimal Current { get; set; }
+        public decimal Voltage { get; set; }
         public bool Is_Valid { get; set; }
-        public double Total { get; set; }
-        public double TotalReturned { get; set; }
+        public decimal Total { get; set; }
+        public decimal TotalReturned { get; set; }
     }
 
     public class ShellyEmeterData
@@ -26,28 +26,28 @@ namespace ShellyConnector.DataContracts
     public class ShellyEmeterGen3PowerData
     {
         public int id { get; set; }
-        public float a_current { get; set; }
-        public float a_voltage { get; set; }
-        public float a_act_power { get; set; }
-        public float a_aprt_power { get; set; }
-        public float a_pf { get; set; }
-        public float a_freq { get; set; }
-        public float b_current { get; set; }
-        public float b_voltage { get; set; }
-        public float b_act_power { get; set; }
-        public float b_aprt_power { get; set; }
-        public float b_pf { get; set; }
-        public float b_freq { get; set; }
-        public float c_current { get; set; }
-        public float c_voltage { get; set; }
-        public float c_act_power { get; set; }
-        public float c_aprt_power { get; set; }
-        public float c_pf { get; set; }
-        public float c_freq { get; set; }
+        public decimal a_current { get; set; }
+        public decimal a_voltage { get; set; }
+        public decimal a_act_power { get; set; }
+        public decimal a_aprt_power { get; set; }
+        public decimal a_pf { get; set; }
+        public decimal a_freq { get; set; }
+        public decimal b_current { get; set; }
+        public decimal b_voltage { get; set; }
+        public decimal b_act_power { get; set; }
+        public decimal b_aprt_power { get; set; }
+        public decimal b_pf { get; set; }
+        public decimal b_freq { get; set; }
+        public decimal c_current { get; set; }
+        public decimal c_voltage { get; set; }
+        public decimal c_act_power { get; set; }
+        public decimal c_aprt_power { get; set; }
+        public decimal c_pf { get; set; }
+        public decimal c_freq { get; set; }
         public object n_current { get; set; }
-        public float total_current { get; set; }
-        public float total_act_power { get; set; }
-        public float total_aprt_power { get; set; }
+        public decimal total_current { get; set; }
+        public decimal total_act_power { get; set; }
+        public decimal total_aprt_power { get; set; }
         public object[] user_calibrated_phase { get; set; }
     }
 
@@ -55,14 +55,14 @@ namespace ShellyConnector.DataContracts
     public class ShellyEmeterGen3EnergyData
     {
         public int id { get; set; }
-        public float a_total_act_energy { get; set; }
-        public float a_total_act_ret_energy { get; set; }
-        public float b_total_act_energy { get; set; }
-        public float b_total_act_ret_energy { get; set; }
-        public float c_total_act_energy { get; set; }
-        public float c_total_act_ret_energy { get; set; }
-        public float total_act { get; set; }
-        public float total_act_ret { get; set; }
+        public decimal a_total_act_energy { get; set; }
+        public decimal a_total_act_ret_energy { get; set; }
+        public decimal b_total_act_energy { get; set; }
+        public decimal b_total_act_ret_energy { get; set; }
+        public decimal c_total_act_energy { get; set; }
+        public decimal c_total_act_ret_energy { get; set; }
+        public decimal total_act { get; set; }
+        public decimal total_act_ret { get; set; }
     }
 
 }
