+++ ShellyConnector/ShellyConnector/DataContracts/ShellyPlugMeterData.cs
@@ -2,11 +2,11 @@
 {
     public class ShellyPlugMeterData
     {
-        public double Power { get; set; }
-        public double Overpower { get; set; }
+        public decimal Power { get; set; }
+        public decimal Overpower { get; set; }
         public bool Is_Valid { get; set; }
         public int Timestamp { get; set; }
-        public List<double> Counters { get; set; }
-        public double Total { get; set; }
+        public List<decimal> Counters { get; set; }
+        public decimal Total { get; set; }
     }
 }
